% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/microeconomics.R
=======
% Please edit documentation in R/supply_and_demand.R
>>>>>>> origin/initial
\name{create_market}
\alias{create_market}
\title{Create a series of elements to analize market's dinamic}
\usage{
<<<<<<< HEAD
create_market(price_q0, slope, perfect_e = NULL, perfect_i = NULL)
=======
create_market(price_q0, slope, perfect_e = NA_real_, perfect_i = NA_real_)
>>>>>>> origin/initial
}
\arguments{
\item{price_q0}{It's the parameter \code{a} and for both, demanded and/or supplied quantities are equal to 0}

<<<<<<< HEAD
\item{slope}{Ir's the parameter \code{b} and it's the rate at witch prices changes due to a change of the quantity demanded or supplied}

\item{perfect_e}{Perfectly \strong{elastic} demand (negative value) or supply (positive value)}

\item{perfect_i}{Perfectly \strong{inelastic} demand (negative value) or supply (positive value)}
}
\value{
Element of class \strong{"market_curves"}. The result is a list with the following elements, for each of
the curves:
\strong{name:} Wether it is a demand or supply curve
\strong{funs:} Functional form of the curve
\strong{equation:} Latex formula to print in
\strong{coeficients:} The value of the slopes and intercepts as matrix
\strong{intercept:} The value of the intercept (plotting porpuses)
=======
\item{slope}{It's the parameter \code{b} and it's the rate at wich prices changes due to a change of the quantity demanded or supplied#'}

\item{perfect_e}{Extreme cases for perfectly elastic demand (-) or supply(+)}

\item{perfect_i}{Extreme cases for perfectly inelastic demand (-) or supply(+)}
}
\value{
A list with all the market needed to construct our market
>>>>>>> origin/initial
}
\description{
Given a set of prices and quantities create all the market needed to
analize a market using linear curves like:
}
\details{
\deqn{$$P = a + bx$$}
}
\examples{

# Make a demand curve:

demand <- create_market(price_q0 = 200,slope = -4)

# Make a supply curve:

supply <- create_market(price_q0 = 100,slope = 2)

# Find the equilibrium:

equiilibrium <- create_market(price_q0 = c(200,100),slope = c(-4,2))

# Many demands and one supply:

equiilibrium <- create_market(price_q0 = c(200,100,150),slope = c(-4,2,3))

# Works the same way for many supply curves
}
