% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{dice_mean_mod}
\alias{dice_mean_mod}
\title{Draw a fair dice \strong{sample_n_t} times
Theoretical mean: 3.5
Theoretical mean: 1.71}
\usage{
dice_mean_mod(sample_n)
}
\arguments{
\item{sample_n}{Draws}
}
\value{
Return the standarized value of the mean draws.
}
\description{
Draw a fair dice \strong{sample_n_t} times
Theoretical mean: 3.5
Theoretical mean: 1.71
}
