% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R
\name{expected_utility}
\alias{expected_utility}
\title{Calculate the expected utility from a bet}
\usage{
expected_utility(
  earnings = c(),
  losings = c(),
  probabilities = c(),
  u_function = function(x) {     x },
  endowment = 0
)
}
\arguments{
\item{earnings}{value you get if you win the bet}

\item{losings}{value you loose if you do not win the bet (always negative numbers)}

\item{probabilities}{set of probabilities for the options of the bet
(the order should be as follows c(earnings,losings) )}

\item{u_function}{Form of the utility function}

\item{endowment}{Initial endowment}
}
\value{
Expected utility from a bet
}
\description{
Calculate the expected utility from a bet
}
