% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R
\name{expected_value}
\alias{expected_value}
\title{Calculate the expected values from a bet}
\usage{
expected_value(
  earnings = c(),
  losings = c(),
  probabilities = c(),
  endowment = 0
)
}
\arguments{
\item{earnings}{value you get if you win the bet}

\item{losings}{value you loose if you do not win the bet}

\item{probabilities}{set of probabilities for the options of the bet
(the order should be as follows c(earnings,losings) )}

\item{endowment}{Initial endowment}
}
\value{
Expected value of a bet
}
\description{
A fair bet is the one which leads to a expected value equal to cero.
Someone who is unwilling to make a fair bet is risk averse. A person
who is indifferent about making a fair bet is risk neutral. A person who
is risk preferring will make a fair bet. (Perloff J. M., 2020)
}
