% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supply_and_demand.R
\name{linear_curve}
\alias{linear_curve}
\title{Plot a demand, supply or both curves for a market}
\usage{
linear_curve(price_q0 = numeric(), slope = numeric())
}
\arguments{
\item{price_q0}{It's the parameter \code{a} and for both, demanded and/or supplied quantities are equal to 0}

\item{slope}{Ir's the parameter \code{b} and it's the rate at wich prices changes due to a change of the quantity demanded or supplied}
}
\value{

}
\description{
Consider the next linear expressiion for a supply (demand is the same especification but with a negative b) curve:
$$P = a + b*q$$
Then:
}
