% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxform.R
\name{taxform_barras_100}
\alias{taxform_barras_100}
\title{Title}
\usage{
taxform_barras_100(
  tabla = tibble(),
  x_var = character(),
  y_var = character(),
  x_title = "",
  y_title = "",
  fill_var = character(),
  title = character(),
  descripcion = character(),
  paleta = "default"
)
}
\arguments{
\item{descripcion}{}
}
\value{

}
\description{
Title
}
