% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated.R
\name{viepi_duplicados}
\alias{viepi_duplicados}
\title{Creación de la tabla de muestras}
\usage{
viepi_duplicados(base_prec = NULL, identificador_unico = character())
}
\arguments{
\item{base_prec}{Base después de precisión}

\item{identificador_unico}{Nombre del identificador único}
}
\value{
Lista con dos bases:
\itemize{
\item \code{1} Para los que tienen muestra
\item \code{0} Para los que no tienen muestra
}
}
\description{
Creación de la tabla de muestras
}
\examples{

viepi_duplicados(viepi_prec)
}
