% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supply_and_demand.R
\name{create_market}
\alias{create_market}
\title{Create a series of elements to analize market's dinamic}
\usage{
create_market(price_q0, slope)
}
\arguments{
\item{price_q0}{It's the parameter \code{a} and for both, demanded and/or supplied quantities are equal to 0}

\item{slope}{Ir's the parameter \code{b} and it's the rate at wich prices changes due to a change of the quantity demanded or supplied#'}
}
\value{
A list with all the market needed to construct our market
}
\description{
Given a set of prices and quantities create all the market needed to
analize a market using linear curves like:
}
\details{
\deqn{$$P = a + bx$$}
}
\examples{

# Make a demand curve:

demand <- create_market(price_q0 = 200,slope = -4)

# Make a supply curve:

supply <- create_market(price_q0 = 100,slope = 2)

# Find the equilibrium:

equiilibrium <- create_market(price_q0 = c(200,100),slope = c(-4,2))

# Many demands and one supply:

equiilibrium <- create_market(price_q0 = c(200,100,150),slope = c(-4,2,3))

# Works the same way for many supply curves
}
