% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microeconomics.R
\name{linear_curve}
\alias{linear_curve}
\title{Plot a demand, supply or both curves for a market}
\usage{
linear_curve(market, market_name = NULL)
}
\arguments{
\item{market}{result from \code{create_market} function}

\item{market_name}{A title for the analyzed market}
}
\value{
Return a list with some elements_
\itemize{
\item \code{curve} PLot of the analyzed market
\item \code{market} all the elements behind the market
\item \code{equilibrium} if at least one curve is a supply curve and at least one is a demand curve then
the equilibriums are calculated
}
}
\description{
Plot a demand, supply or both curves for a market
}
\examples{

# market <- create_market(price_q0 = c(100,200),slope = c(-1,1))
# linear_curve(market = market,market_name = "Black market")
}
